Move phantom logic to hook
