+++ components/ConnectPhantomButton.tsx
@@ -23,133 +23,14 @@ import {
   Transaction,
 } from "@solana/web3.js";
 import { RPC_CONNECTION } from "../constants/Solana";
+import { usePhantom } from "../hooks/usePhantom";
+
 const onConnectRedirectLink = Linking.createURL("onConnect");
-
-const buildUrl = (path: string, params: URLSearchParams) =>
-  `https://phantom.app/ul/v1/${path}?${params.toString()}`;
-
-const decryptPayload = (
-  data: string,
-  nonce: string,
-  sharedSecret?: Uint8Array
-) => {
-  if (!sharedSecret) throw new Error("missing shared secret");
-
-  const decryptedData = nacl.box.open.after(
-    bs58.decode(data),
-    bs58.decode(nonce),
-    sharedSecret
-  );
-  if (!decryptedData) {
-    throw new Error("Unable to decrypt data");
-  }
-  return JSON.parse(Buffer.from(decryptedData).toString("utf8"));
-};
-
-const encryptPayload = (payload: any, sharedSecret?: Uint8Array) => {
-  if (!sharedSecret) throw new Error("missing shared secret");
-
-  const nonce = nacl.randomBytes(24);
-
-  const encryptedPayload = nacl.box.after(
-    Buffer.from(JSON.stringify(payload)),
-    nonce,
-    sharedSecret
-  );
-
-  return [nonce, encryptedPayload];
-};
 
 interface ConnectWalletProps {}
 
 export const ConnectPhantomButton = ({}: ConnectWalletProps) => {
-  const theme = useTheme();
-  const dispatch = useAppDispatch();
-  const { publicKey } = useAppSelector((state) => state.wallet);
-  const [deepLink, setDeepLink] = useState<string>("");
-  const connection = new Connection(RPC_CONNECTION);
-  const scrollViewRef = useRef<any>(null);
-
-  // store dappKeyPair, sharedSecret, session and account SECURELY on device
-  // to avoid having to reconnect users.
-  const [dappKeyPair] = useState(nacl.box.keyPair());
-  const [sharedSecret, setSharedSecret] = useState<Uint8Array>();
-  const [session, setSession] = useState<string>();
-  const [phantomWalletPublicKey, setPhantomWalletPublicKey] =
-    useState<PublicKey>();
-
-  useEffect(() => {
-    (async () => {
-      const initialUrl = await Linking.getInitialURL();
-      if (initialUrl) {
-        setDeepLink(initialUrl);
-      }
-    })();
-    Linking.addEventListener("url", handleDeepLink);
-    return () => {
-      Linking.removeEventListener("url", handleDeepLink);
-    };
-  }, []);
-
-  const handleDeepLink = ({ url }: Linking.EventType) => {
-    setDeepLink(url);
-  };
-
-  useEffect(() => {
-    if (!deepLink) return;
-
-    const url = new URL(deepLink);
-    const params = url.searchParams;
-
-    if (params.get("errorCode")) {
-      return;
-    }
-
-    if (/onConnect/.test(url.pathname)) {
-      const sharedSecretDapp = nacl.box.before(
-        bs58.decode(params.get("phantom_encryption_public_key")!),
-        dappKeyPair.secretKey
-      );
-
-      const connectData = decryptPayload(
-        params.get("data")!,
-        params.get("nonce")!,
-        sharedSecretDapp
-      );
-
-      setSharedSecret(sharedSecretDapp);
-      setSession(connectData.session);
-      setPhantomWalletPublicKey(new PublicKey(connectData.public_key));
-
-      const jsonValue = JSON.stringify({
-        publicKey: connectData.public_key,
-        userInfo: {},
-      });
-
-      AsyncStorage.setItem("@walletInfo", jsonValue);
-      dispatch(
-        setWallet({
-          publicKey: connectData.public_key,
-          privateKey: "",
-          userInfo: {},
-          walletType: "web3auth",
-        })
-      );
-    } else if (/onDisconnect/.test(url.pathname)) {
-    }
-  }, [deepLink]);
-
-  const connect = async () => {
-    const params = new URLSearchParams({
-      dapp_encryption_public_key: bs58.encode(dappKeyPair.publicKey),
-      cluster: "mainnet-beta",
-      app_url: "https://phantom.app",
-      redirect_link: onConnectRedirectLink,
-    });
-
-    const url = buildUrl("connect", params);
-    Linking.openURL(url);
-  };
+  const { connect } = usePhantom();
 
   return (
     <ConnectButton onPress={() => connect()}>
